requirements.txt
aiofiles==0.8.0
azure-cognitiveservices-speech==1.27.0
ffmpeg-python==0.2.0
pytube==12.1.3
pydub==0.25.1
librosa==0.10.0